+++ up_fast_downward/fast_downward.py
@@ -32,7 +32,7 @@ class FastDownwardPDDLPlannerBase(PDDLPlanner):
         fast_downward_search_time_limit: Optional[str] = None,
         log_level: str = "info",
     ):
-        super().__init__()
+        super().__init__(rewrite_bool_assignments=True)
         self._fd_alias = fast_downward_alias
         self._fd_search_config = fast_downward_search_config
         self._fd_anytime_alias = fast_downward_anytime_alias
@@ -260,9 +260,12 @@ class FastDownwardPDDLPlanner(FastDownwardPDDLPlannerBase, mixins.AnytimePlanner
         supported_kind.set_conditions_kind("DISJUNCTIVE_CONDITIONS")
         supported_kind.set_conditions_kind("EXISTENTIAL_CONDITIONS")
         supported_kind.set_conditions_kind("UNIVERSAL_CONDITIONS")
-        supported_kind.set_conditions_kind("EQUALITY")
+        supported_kind.set_conditions_kind("EQUALITIES")
         supported_kind.set_effects_kind("CONDITIONAL_EFFECTS")
+        supported_kind.set_effects_kind("STATIC_FLUENTS_IN_BOOLEAN_ASSIGNMENTS")
+        supported_kind.set_effects_kind("FLUENTS_IN_BOOLEAN_ASSIGNMENTS")
         supported_kind.set_quality_metrics("ACTIONS_COST")
+        supported_kind.set_actions_cost_kind("STATIC_FLUENTS_IN_ACTIONS_COST")
         supported_kind.set_quality_metrics("PLAN_LENGTH")
         return supported_kind
 
@@ -310,8 +313,9 @@ class FastDownwardOptimalPDDLPlanner(FastDownwardPDDLPlannerBase):
         supported_kind.set_conditions_kind("DISJUNCTIVE_CONDITIONS")
         supported_kind.set_conditions_kind("EXISTENTIAL_CONDITIONS")
         supported_kind.set_conditions_kind("UNIVERSAL_CONDITIONS")
-        supported_kind.set_conditions_kind("EQUALITY")
+        supported_kind.set_conditions_kind("EQUALITIES")
         supported_kind.set_quality_metrics("ACTIONS_COST")
+        supported_kind.set_actions_cost_kind("STATIC_FLUENTS_IN_ACTIONS_COST")
         supported_kind.set_quality_metrics("PLAN_LENGTH")
         return supported_kind